fix: in filter with key-config
