+++ src/test/conversion-union.test.ts
@@ -10,28 +10,28 @@ interface Data {
 }
 
 test('it should convert a primitive of a union type', () => {
-  const simpleConfig: FilterSetConfig<Data> = {
+  const config: FilterSetConfig<Data> = {
     number: {lt: 123},
   }
-  const converted = convertFilterSetConfig(simpleConfig)
+  const converted = convertFilterSetConfig(config)
   // eslint-disable-next-line camelcase
   expect(converted).toEqual({number__lt: 123})
 })
 
 test('it should convert a complex type of a union type', () => {
   const special = {internal: 123}
-  const simpleConfig: FilterSetConfig<Data> = {
+  const config: FilterSetConfig<Data> = {
     number: {value: special},
   }
-  const converted = convertFilterSetConfig(simpleConfig)
+  const converted = convertFilterSetConfig(config)
   expect(converted).toEqual({number: special})
 })
 
 test('the config should be able to filter attributes of complex types', () => {
-  const simpleConfig: FilterSetConfig<Data> = {
-    number: {internal: {value: 123}},
+  const config: FilterSetConfig<Data> = {
+    number: {internal: {lt: 123}},
   }
-  const converted = convertFilterSetConfig(simpleConfig)
+  const converted = convertFilterSetConfig(config)
   // eslint-disable-next-line camelcase
-  expect(converted).toEqual({number__internal: 123})
+  expect(converted).toEqual({number__internal__lt: 123})
 })
